Code/requirements.txt
# Core framework
langchain>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.0
openai>=1.30.0
anthropic>=0.28.0

# Vector store / embeddings
chromadb>=0.5.0
sentence-transformers>=3.0.0
faiss-cpu>=1.8.0

# IaC security validators
checkov>=3.2.0

# Code analysis
transformers>=4.40.0
torch>=2.2.0

# Scraping
aiohttp>=3.9.0

# Utilities
pyyaml>=6.0
python-dotenv>=1.0.0
rich>=13.7.0
typer>=0.12.0
pydantic>=2.7.0

# Testing
pytest>=8.2.0
pytest-cov>=5.0.0
